Crate sp_domains_fraud_proof

Subspace fraud proof primitives for consensus chain.

Modules

execution_prover

This module provides the feature of generating and verifying the execution proof used in the Subspace fraud proof mechanism.

fraud_proof

fraud_proof_runtime_interface

Domain fraud proof related runtime interface

storage_proof

verification

Structs

DomainInherentExtrinsic

DomainInherentExtrinsicData

FraudProofExtension

Domains fraud proof host function

FraudProofHostFunctionsImpl

Trait Impl to query and verify Domains Fraud proof.

HostFunctions

Provides implementations for the extern host functions.

Enums

DomainChainAllowlistUpdateExtrinsic

Type that maybe holds an encoded update domain chain allowlist extrinsic

DomainStorageKeyRequest

InvalidTransactionCode

Custom invalid validity code for the extrinsics in pallet-domains.

SetCodeExtrinsic

Type that maybe holds an encoded set_code extrinsic with upgraded runtime

StatelessDomainRuntimeCall

StorageKeyRequest

Type that specifies the request of storage keys

Traits

FraudProofApi

API necessary for fraud proof.

FraudProofHostFunctions

Trait to query and verify Domains Fraud proof.