sp_domains_fraud_proof
Module fraud_proof

Structs

DomainRuntimeCodeAt

FraudProof

InvalidBlockFeesProof

Represents an invalid block fees proof.

InvalidBundlesProof

InvalidDomainBlockHashProof

Represents an invalid domain block hash fraud proof.

InvalidExtrinsicsRootProof

InvalidStateTransitionProof

Proves an invalid state transition by challenging the trace at specific index in a bad receipt.

InvalidTransfersProof

Represents an invalid transfers proof.

MmrRootProof

ValidBundleDigest

Represents a valid bundle index and all the extrinsics within that bundle.

ValidBundleProof

Fraud proof for the valid bundles in ExecutionReceipt::inboxed_bundles

Enums

ApplyExtrinsicMismatch

Mismatch type possible for ApplyExtrinsic execution phase

ExecutionPhase

A phase of a block’s execution, carrying necessary information needed for verifying the invalid state transition proof.

FinalizeBlockMismatch

Mismatch type possible for FinalizBlock execution phase

FraudProofVariant

InvalidBundlesProofData

VerificationError

Error type of fraud proof verification on consensus node.