Enum sp_domains_fraud_proof::fraud_proof::VerificationError

pub enum VerificationError<DomainHash> {

    ConsensusBlockHashNotFound,
    BadExecutionProof,
    InvalidProof,
    InitializeBlockOrApplyExtrinsicDecode(Error),
    StorageRootDecode(Error),
    HeaderDecode(Error),
    ValidTransaction,
    StateNotFound(Vec<u8>),
    Decode(Error),
    RuntimeApi(ApiError),
    Client(Error),
    InvalidStorageProof,
    SignerNotFound,
    DomainStateRootNotFound,
    RuntimeCode(String),
    FailedToGetDomainRuntimeCode,
    FailedToGetDomainTransfersStorageKey,
    Oneshot(String),
    InvalidExecutionTrace,
    InvalidApplyExtrinsicTraceIndex,
    InvalidLongerMismatchTraceIndex,
    InvalidApplyExtrinsicCallData,
    InvalidBundleDigest,
    FailedToGetBlockRandomness,
    FailedToDeriveDomainTimestampExtrinsic,
    FailedToDeriveConsensusChainByteFeeExtrinsic,
    FailedToDeriveDomainSetCodeExtrinsic,
    BundleNotFound,
    UnexpectedTargetedBundleEntry {
        bundle_index: u32,
        fraud_proof_invalid_type_of_proof: InvalidBundleType,
        targeted_entry_bundle: BundleValidity<DomainHash>,
    },
    FailedToGetResponseFromTxRangeHostFn,
    FailedToGetDomainBundleBody,
    FailedToDeriveBundleDigest,
    TargetValidBundleNotFound,
    FailedToCheckInherentExtrinsic,
    FailedToCheckExtrinsicDecodable,
    InvalidBundleEquivocationFraudProof(InvalidBundleEquivocationError),
    FailedToCheckExtrinsicsInSingleContext,

}

Error type of fraud proof verification on consensus node.

Variants

ConsensusBlockHashNotFound

Hash of the consensus block being challenged not found.

BadExecutionProof

Failed to pass the execution proof check.

InvalidProof

The fraud proof prove nothing invalid

InitializeBlockOrApplyExtrinsicDecode(Error)

Failed to decode the return value of initialize_block and apply_extrinsic.

StorageRootDecode(Error)

Failed to decode the storage root produced by verifying initialize_block or apply_extrinsic.

HeaderDecode(Error)

Failed to decode the header produced by finalize_block.

ValidTransaction

Transaction validity check passes.

StateNotFound(Vec<u8>)

State not found in the storage proof.

Decode(Error)

Decode error.

RuntimeApi(ApiError)

Runtime api error.

Client(Error)

Runtime api error.

InvalidStorageProof

Invalid storage proof.

SignerNotFound

Can not find signer from the domain extrinsic.

DomainStateRootNotFound

Domain state root not found.

RuntimeCode(String)

Fail to get runtime code.

FailedToGetDomainRuntimeCode

FailedToGetDomainTransfersStorageKey

Oneshot(String)

InvalidExecutionTrace

InvalidApplyExtrinsicTraceIndex

InvalidLongerMismatchTraceIndex

InvalidApplyExtrinsicCallData

InvalidBundleDigest

Invalid bundle digest

FailedToGetBlockRandomness

Failed to get block randomness

FailedToDeriveDomainTimestampExtrinsic

Failed to derive domain timestamp extrinsic

FailedToDeriveConsensusChainByteFeeExtrinsic

Failed to derive consensus chain byte fee extrinsic

FailedToDeriveDomainSetCodeExtrinsic

Failed to derive domain set code extrinsic

BundleNotFound

Bundle with requested index not found in execution receipt

UnexpectedTargetedBundleEntry

Fields

bundle_index: u32

fraud_proof_invalid_type_of_proof: InvalidBundleType

targeted_entry_bundle: BundleValidity<DomainHash>

Invalid bundle entry in bad receipt was expected to be valid but instead found invalid entry

FailedToGetResponseFromTxRangeHostFn

Tx range host function did not return response (returned None)

FailedToGetDomainBundleBody

Failed to get the bundle body

FailedToDeriveBundleDigest

Failed to derive bundle digest

TargetValidBundleNotFound

The target valid bundle not found from the target bad receipt

FailedToCheckInherentExtrinsic

Failed to check if a given extrinsic is inherent or not.

FailedToCheckExtrinsicDecodable

Failed to check if a given extrinsic is decodable or not.

InvalidBundleEquivocationFraudProof(InvalidBundleEquivocationError)

Invalid bundle equivocation fraud proof.

FailedToCheckExtrinsicsInSingleContext

Failed to check extrinsics in single context