sp_domains_fraud_proof

Module fraud_proof_runtime_interface

Domain fraud proof related runtime interface

Structs

HostFunctions

Provides implementations for the extern host functions.

Functions

bundle_weight

check_extrinsics_in_single_context

construct_domain_inherent_extrinsic

derive_bundle_digest

Derive the bundle digest for the given bundle body.

domain_runtime_call

domain_storage_key

execution_proof_check

Check the execution proof with also included domain block id.

extract_xdm_mmr_proof