Module sp_domains_fraud_proof::storage_proof

Structs

BlockDigestProof

BlockRandomnessProof

ConsensusTransactionByteFeeProof

DomainChainsAllowlistUpdateStorageProof

DomainInherentExtrinsicDataProof

DomainRuntimeCodeProof

DomainSudoCallStorageProof

DynamicCostOfStorageProof

MaybeDomainRuntimeUpgradedProof

MmrRootStorageProof

OpaqueBundleWithProof

SuccessfulBundlesProof

TimestampStorageProof

Enums

FraudProofStorageKeyRequest

GenerationError

VerificationError

Traits

BasicStorageProof

FraudProofStorageKeyProvider

Trait to get storage keys in the runtime i.e. when verifying the storage proof

FraudProofStorageKeyProviderInstance

Trait to get storage keys in the client i.e. when generating the storage proof