Crate sp_domains_fraud_proof

Subspace fraud proof primitives for consensus chain.

Modules

bundle_equivocation

Module to check bundle equivocation and produce the Equivocation fraud proof. This is mostly derived from the sc_consensus_slots::aux_schema with changes adapted for Bundle headers instead of block headers

execution_prover

This module provides the feature of generating and verifying the execution proof used in the Subspace fraud proof mechanism. The execution is more fine-grained than the entire block execution, block execution hooks (initialize_block and finalize_block) and any specific extrinsic execution are supported.

fraud_proof

fraud_proof_runtime_interface

Domain fraud proof related runtime interface

verification

Structs

FraudProofExtension

Domains fraud proof host function

FraudProofHostFunctionsImpl

Trait Impl to query and verify Domains Fraud proof.

HostFunctions

Provides implementations for the extern host functions.

Enums

FraudProofVerificationInfoRequest

Request type to fetch required verification information for fraud proof through Host function.

FraudProofVerificationInfoResponse

Response holds required verification information for fraud proof from Host function.

InvalidTransactionCode

Custom invalid validity code for the extrinsics in pallet-domains.

SetCodeExtrinsic

Type that maybe holds an encoded set_code extrinsic with upgraded runtime

StorageKeyRequest

Type that specifies the request of storage keys

Traits

FraudProofApi

API necessary for fraud proof.

FraudProofHostFunctions

Trait to query and verify Domains Fraud proof.