sp_domains_fraud_proof

Module fraud_proof

Structs

DomainRuntimeCodeAt

FraudProof

Fraud proof for domains.

InvalidBlockFeesProof

Represents an invalid block fees proof.

InvalidBundlesProof

A proof about a bundle that was marked invalid (but might or might not actually be invalid).

InvalidDomainBlockHashProof

Represents an invalid domain block hash fraud proof.

InvalidExtrinsicsRootProof

InvalidStateTransitionProof

Proves an invalid state transition by challenging the trace at specific index in a bad receipt.

InvalidTransfersProof

Represents an invalid transfers proof.

MmrRootProof

ValidBundleDigest

Represents a valid bundle index and all the extrinsics within that bundle.

ValidBundleProof

Fraud proof for the valid bundles in ExecutionReceipt::inboxed_bundles

Enums

ApplyExtrinsicMismatch

Mismatch type possible for ApplyExtrinsic execution phase

ExecutionPhase

A phase of a block’s execution, carrying necessary information needed for verifying the invalid state transition proof.

FinalizeBlockMismatch

Mismatch type possible for FinalizBlock execution phase

FraudProofVariant

InvalidBundlesProofData

Invalid versioned bundle proof data.

VerificationError

Error type of fraud proof verification on consensus node.