Module sp_domains_fraud_proof::fraud_proof

Structs

BundleEquivocationProof

Represents a bundle equivocation proof. An equivocation happens when an executor produces more than one bundle on the same slot. The proof of equivocation are the given distinct bundle headers that were signed by the validator and which include the slot number.

ImproperTransactionSortitionProof

Represents an invalid transaction proof.

InvalidBlockFeesProof

Represents an invalid block fees proof.

InvalidBundlesFraudProof

InvalidDomainBlockHashProof

Represents an invalid domain block hash fraud proof.

InvalidExtrinsicsRootProof

Represents an Invalid domain extrinsics root proof with necessary info for verification.

InvalidStateTransitionProof

Proves an invalid state transition by challenging the trace at specific index in a bad receipt.

InvalidTransactionProof

Represents an invalid transaction proof.

InvalidTransfersProof

Represents an invalid transfers proof.

ValidBundleDigest

Represents a valid bundle index and all the extrinsics within that bundle.

ValidBundleProof

Fraud proof for the valid bundles in ExecutionReceipt::inboxed_bundles

Enums

ApplyExtrinsicMismatch

Mismatch type possible for ApplyExtrinsic execution phase

ExecutionPhase

A phase of a block’s execution, carrying necessary information needed for verifying the invalid state transition proof.

FinalizeBlockMismatch

Mismatch type possible for FinalizBlock execution phase

FraudProof

Fraud proof.

VerificationError

Error type of fraud proof verification on consensus node.

Functions

dummy_invalid_state_transition_proof

operator_block_fees_final_key

This is a representation of actual Block Fees storage in pallet-block-fees. Any change in key or value there should be changed here accordingly.

system_digest_final_key

Digest storage key in frame_system. Unfortunately, the digest storage is private and not possible to derive the key from it directly.