Module sp_domains_fraud_proof::fraud_proof_runtime_interface

Domain fraud proof related runtime interface

Structs

HostFunctions

Provides implementations for the extern host functions.

Functions

bundle_weight

check_extrinsics_in_single_context

construct_domain_inherent_extrinsic

derive_bundle_digest

Derive the bundle digest for the given bundle body.

domain_runtime_call

domain_storage_key

execution_proof_check

Check the execution proof with also included domain block id.

extract_xdm_mmr_proof

get_fraud_proof_verification_info

Returns required fraud proof verification information to the runtime through host function.