sp_domains_fraud_proof

Module storage_proof

Structs

DomainChainsAllowlistUpdateStorageProof

DomainRuntimeCodeProof

DomainRuntimeUpgradesProof

DomainSudoCallStorageProof

EvmDomainContractCreationAllowedByCallStorageProof

InherentExtrinsicData

InvalidInherentExtrinsicDataProof

MaybeDomainRuntimeUpgradedProof

A proof of a single domain runtime upgrade (or that there wasn’t an upgrade).

MmrRootStorageProof

OpaqueBundleWithProof

Bundle with proof data for fraud proof.

SuccessfulBundlesProof

Enums

FraudProofStorageKeyRequest

GenerationError

VerificationError

Traits

BasicStorageProof

FraudProofStorageKeyProvider

Trait to get storage keys in the runtime i.e. when verifying the storage proof

FraudProofStorageKeyProviderInstance

Trait to get storage keys in the client i.e. when generating the storage proof